remove/refactor more unused utilities
