+++ src/fairchem/core/scripts/fit_normalizers.py
@@ -11,9 +11,10 @@ import argparse
 import logging
 from pathlib import Path
 
+from fairchem.experimental.legacy.utils import load_config
+
 from fairchem.core.common.registry import registry
-from fairchem.core.common.utils import load_config
-from fairchem.core.modules.normalization._load_utils import save_checkpoint
+from fairchem.core.common.utils import save_checkpoint
 from fairchem.core.modules.normalization.element_references import (
     create_element_references,
 )
