+++ src/fairchem/core/scripts/fit_references.py
@@ -11,8 +11,9 @@ import argparse
 import logging
 from pathlib import Path
 
+from fairchem.experimental.legacy.utils import load_config
+
 from fairchem.core.common.registry import registry
-from fairchem.core.common.utils import load_config
 from fairchem.core.modules.normalization._load_utils import save_checkpoint
 from fairchem.core.modules.normalization.element_references import fit_linear_references
